restructured UTs of quota-app, reports and scheduler
